Removed the old stuff, added new Code
You still can look back at the old code by following the changed I did to the project. That old code though, is not necessary anymore, as we got the whole entity list now.

This new project I added to the source folder contains the simplest form of a healer bot.